+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Log a debug message when a job was added to an undefined queue
 
 ## [0.3.1] - 2017-01-11
 ### Fixed